refactor

+++ rust/src/lib.rs
@@ -9,7 +9,7 @@ impl Game {
     pub fn roll(&mut self, pins : u8) {
     }
 
-    pub fn score(&mut self) -> i32 {
+    pub fn score(&mut self) -> u16 {
         0
     }
 }
